termination msg in PGPROC
This commit adds message in PGPROC, that can be set
before terminating/cancelling connection, in order to add
this message when backend is responding to client with
'FATAL: terminating connection..' or 'ERROR: cancelling statement..'

For now there is no separated functions to perform termination with msg,
but that functions can be simply added in extension. This requires
implementation of pg_terminate_backend/pg_cancel_backend to be accessible
from extenions, so pg_terminate_backend_impl and pg_cancel_backend_impl was
introduced as public functions.

Signed-off-by: roman khapov <r.khapov@ya.ru>